+++ lib/src/constants/message_constants.dart
@@ -1,6 +1,5 @@
 /// Stores all the messages used throughout the app that communicates
 /// with the user of the package.
-
 /// Message shown when we encounter a failure during generation that's caused by an invalid
 /// app structure.
 const String kInvalidDolphinStructureAppFile =