+++ lib/src/exceptions/config_file_not_found_exception.dart
@@ -1,3 +1,6 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+@internal
 class ConfigFileNotFoundException implements Exception {
   final String message;
   final bool shouldHaltCommand;