+++ lib/src/exceptions/invalid_dolphin_structure_exception.dart
@@ -1,3 +1,6 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+@internal
 class InvalidDolphinStructureException implements Exception {
   final String message;
   InvalidDolphinStructureException(this.message);