+++ lib/src/templates/compiled_constants.dart
@@ -1,6 +1,6 @@
-/// NOTE: This is generated code from the compileTemplates command. Do not 
-///       modify by hand.
-///       This file should be checked into source control.
+// NOTE: This is generated code from the compileTemplates command. Do not
+//       modify by hand.
+//       This file should be checked into source control.
 
 Map<String, List<String>> kCompiledTemplateTypes = {
   'app': [